incremental PASSESTOKILL; added in WaveController; currently increases by 0.5; caps at 8

+++ Assets/Scripts/Controller Scripts/ComponentGenerator.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using System.Collections;
 
 public class ComponentGenerator {
-	public int PASSESTOKILL = 2;
+	public float PASSESTOKILL = 2;
 	private static ComponentGenerator _cgen;	// Creating a Singleton
 	private TowerComponent[][] _towerParts;
 	private Object[] _bases;
